% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/revision.R
\name{cargar_emg}
\alias{cargar_emg}
\title{Cargar señales EMG desde archivo .csv}
\usage{
emg<-cargar_emg(nombre="Sujeto",n=2,carpetas=c("mov","max"),extension=".csv",sep=",",header=T)}
\description{
1.- Cargar todos los archivos dentro de una o más carpeta. La extnesión puede ser .txt, la separación de las columnas está señalada con el argumento "sep"
\examples{
}
\keyword{cargar señales EMG}
\keyword{biosignalEMG}


