% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/revision.R
\name{processEMG}
\alias{processEMG}
\title{Funciones para análisis de señales EMG}
\usage{
emg<-cargar_emg(nombre="Sujeto",n=2,carpetas=c("mov","max"),extension=".csv",sep=",",header=T)}
\description{
Pasos para el análisis EMG
1.- Cargar todos los archivos. Ver la configuración, carpetas para mov, base, max

2.- Convertir a emg: emg<-llaply_emg(x,samplingrate=1925.93)

3.- Remover offset: emg.dc<-llaply_pr(emg,FUN=dcbiasremoval)

4.- Rectificar: emg.r<-llaply_pr(emg.dc,FUN=rectification,type="fullwave")

5.- Media Móvil: emg.ma<-llaply_pr(emg.r,FUN=movingaverage,wsize=50)

6.- Normalizar:
1º emg.max<-llaply_max(emg.ma)
2º emg.n<-llaply_norm(emg.ma,max=emg.max)

7.- On/off
}
\examples{
}
\keyword{procesamiento EMG}

